requirements.txt
langchain
langchain-community
langchain_groq
chromadb
pandas
python-dotenv
streamlit
sentence-transformers
langchain_huggingface
langchain_chroma